refactor(api-keys): list via server fn, tighten types, doc fixes
Address review feedback:
- List keys through a `listApiKeys` server fn so the active org comes from the
  authenticated context — drops the extra client `getSession()` round-trip and
  the loose `unwrapKeys` shape-probing in favor of a typed response.
- Remove the `unknown` double-cast in `useCreateApiKey`; return the server fn's
  typed result directly, and drop the now-redundant null guard in the dialog's
  onSuccess (the server throws on a missing key).
- Clear the "Copied" reset timer on unmount via a ref + effect.
- `describeApiKeyScopes` preserves declaration order instead of sorting, so
  table badges match the create dialog's order.
- Docs: call it the "Capabilities column" to match the UI (was "Scopes").
